refactor App.js,header,lunchpad
